+++ NEWS.md
@@ -1,6 +1,7 @@
 # usethis 1.2.0.9000
 
 * `create_package()` and `create_project()` return a normalized path, even if target directory does not pre-exist (#227, #228).
+* Moves tidyverse templates to `.github` subdirectory (#165 @batpigandme)
 
 # usethis 1.2.0